+++ uistate-circuit/src/nonWebTest/kotlin/io/github/solcott/uistate/circuit/ProduceContentStateTest.kt
@@ -134,8 +134,7 @@ class ProduceContentStateTest {
     }
   }
 
-  // --- Flow<P>.produceRetainedContentState
-  // ----------------------------------------------------------------------
+  // --- Flow<P>.produceRetainedContentState ---
 
   @Test
   fun aNewParameterFlagsLoadingWithoutDroppingTheCurrentContent() = runTest {